% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibPlot.R
\name{plot.zibSim}
\alias{plot.zibSim}
\title{Visualise a zero-inflated beta regression model for simulated data}
\usage{
\method{plot}{zibSim}(x, ..., microbe_name = "Microbe")
}
\arguments{
\item{x}{An object of class \code{"zibSim"}.}

\item{...}{Catches unused arguments to \code{plot} (not currently
implemented).}

\item{microbe_name}{Name of simulated microbe to title the plot; "Microbe" by
default.}
}
\description{
Plots a simulated zero-inflated beta regression dataset of class
\code{"zibSim"}, separately displaying the zero abundance in the dataset and
fitting a beta regression model to the distribution of the non-zero data,
which is displayed on a histogram.
}
\examples{
n = 1000
t = 3
X <- as.matrix(c(rep(0, n/2 * t), rep(1, n/2 * t)))

y1 <- zibSim(n = n, t = t, a = 0.5, b = 0.8, sigma1 = 2.5, sigma2 = 0.5,
phi = 7.2, alpha = 0.5, beta = -0.5, X = X, Z = X, seed = 6874)

plot(y1, microbe_name = "Simulated Data")

}
\seealso{
\code{\link{plot}}, \code{\link{zibSim}}
}
