% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibClean.R
\name{zibClean}
\alias{zibClean}
\title{Prepare longitudinal microbiome composition data for modelling}
\usage{
zibClean(
  data,
  metadata,
  log_covs,
  beta_covs = log_covs,
  id_column,
  subject_column,
  time_column
)
}
\arguments{
\item{data}{Response variable (relative abundance) for the regression model;
row names must equal the \code{id_column} vector and column names must
equal the names of the different microbes to analyse.}

\item{metadata}{Metadata including information for each individual sample on
all covariates to be included in the model.}

\item{log_covs}{Names of the covariates to be included in the log part of the
model - should equal names of columns in the \code{metadata} data set.}

\item{beta_covs}{Names of the covariates to be included in the beta part of
the model - should equal names of columns in the \code{metadata} data set.
Equal to the \code{log_covs} by default.}

\item{id_column}{Name of the column in \code{metadata} corresponding to the
ID of each sample; should correspond to the rownames of \code{data}; should
be inputted as a character string.}

\item{subject_column}{Name of the column in \code{metadata} corresponding to
the subject each sample is associated with; should be inputted as a
character string.}

\item{time_column}{Name of the column in \code{metadata} corresponding to the
time point each sample was taken at; should be inputted as a character
string.}
}
\value{
An object of class \code{"zibData"}, which is a list containing the
following items;
 \item{\code{rel_abundance}: }{Relative abundance of each species for each sample,
 of class \code{data.frame}.}.
 \item{\code{log_covariates}: }{Covariate matrix X used to simulate the data
 (logistic part), of class \code{data.frame}.}.
 \item{\code{beta_covariates}: }{Covariate matrix Z used to simulate the data
 (beta part), of class \code{data.frame}.}
 \item{\code{spec_names}: }{Vector of the unique species names to be modelled}.
 \item{\code{subject_ind}: }{Ordered index of subject ID's.}
 \item{\code{time_ind}: }{Ordered index of time points for each data point.}
 \item{\code{sample_ind}: }{Ordered index of sample ID's.}
}
\description{
Given two data sets, one detailing the relative abundances of different
microbes for different samples (\code{data}) and one providing information
on the covariates for each individual sample (\code{metadata}), this function
tidies the data up into a \code{"zibData"} object, suitable for fitting
zero-inflated beta regression models to the data using the
\code{\link{zibFitter}} package.
}
\examples{
library(zibFit)
# simulate zibr data and plot

# initialise covariate matrix, simulating a clinical study where half recieve
# the treatment (1) and half do not (0)
n = 1000
t = 3
X <- as.matrix(c(rep(0, n/2 * t), rep(1, n/2 * t)))

y1 <- zibSim(n = n, t = t, a = 0.5, b = 0.8, sigma1 = 2.5, sigma2 = 0.5,
phi = 7.2, alpha = 0.5, beta = -0.5, X = X, Z = X, seed = 6874)

# fit a zibr model to longitudinal microbiome composition data
# simulate five different microbiomes
y2 <- zibSim(n = n, t = t, a = 0.5, b = -0.5, sigma1 = 0.8, sigma2 = 1.2,
phi = 5, alpha = -0.5, beta = -0.5, X = X, Z = X, seed = 6874)
y3 <- zibSim(n = n, t = t, a = 0.2, b = 0.3, sigma1 = 3, sigma2 = 2,
phi = 4, alpha = -0.5, beta = -0.5, X = X, Z = X, seed = 6874)
y4 <- zibSim(n = n, t = t, a = 0.1, b = 0.9, sigma1 = 0.4, sigma2 = 0.5,
phi = 8.1, alpha = 0.5, beta = 0.5, X = X, Z = X, seed = 6874)
y5 <- zibSim(n = n, t = t, a = 0.2, b = 1, sigma1 = 2, sigma2 = 0.5,
phi = 3.4, alpha = -0.5, beta = 0.5, X = X, Z = X, seed = 6874)

ra <- cbind(y1$rel_abundance, y2$rel_abundance, y3$rel_abundance,
y4$rel_abundance, y5$rel_abundance)
rownames(ra) <- rep(1:3000) # row names and column names are required
colnames(ra) <- rep(1:5)

# create covariate data frame
sample_id <- rep(1:3000)
covariates <- data.frame(subject = y1$subject_ind, time = y1$time_ind,
treat = y1$log_covariates$X, sample = sample_id)

zibData = zibClean(data = ra,
metadata = covariates,
log_covs = c("subject", "time", "treat"),
id_column = "sample",
subject_column = "subject",
time_column = "time")

}
\seealso{
\code{zibFitter}, \code{zibSim}
}
