% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibSim.R
\name{zibSim}
\alias{zibSim}
\title{Simulate zero-inflated beta regression models}
\usage{
zibSim(n, t, a, b, sigma1, sigma2, phi, alpha, beta, X, Z, seed = NULL)
}
\arguments{
\item{n}{Number of samples to simulate}

\item{t}{Number of time points to simulate for each sample}

\item{a}{Random effect mean (logistic part)}

\item{b}{Random effect mean (beta part)}

\item{sigma1}{Random effect variance (logistic part); must be a positive
integer}

\item{sigma2}{Random effect variance (beta part); must be a positive integer}

\item{phi}{Dispersion parameter for the beta part}

\item{alpha}{Regression coeffiecients (logistic part); must be the same
length as the number of columns in covariate X}

\item{beta}{Regression coeffiecients (beta part); must be the same length as
the number of columns in covariate Z}

\item{X}{Covariate matrix (logistic part); must be the same length as the
total number of observations, \code{n*t}}

\item{Z}{Covariate matrix (beta part); must be the same length as the total
number of observations, \code{n*t}}

\item{seed}{Simulation seed to be used}
}
\value{
An object of class \code{"zibSim"}, which is a list containing the
following items;
 \item{\code{rel_abundance}: }{Simulated data from the zero-inflated beta regression
 model, using the given parameters}
 \item{\code{log_covariates}: }{Covariate matrix X used to simulate the data
 (logistic part)}
 \item{\code{beta_covariates}: }{Covariate matrix Z used to simulate the data
 (beta part)}
 \item{\code{subject_ind}: }{Ordered index of subject ID's}
 \item{\code{time_ind}: }{Ordered index of time points for each data point}
 \item{\code{sample_ind}: }{Ordered index of sample ID's}
 \item{\code{theta}: }{Named list of all parameters used in the simulations}
}
\description{
Simulates data from a zero-inflated beta regression model, in accordance with
user-defined input parameters. The zero-inflated beta regression model is
commonly used to fit longitudinal microbiome composition data. The model
is specified as follows;
\deqn{
y_{it} \sim
\begin{cases}
0 & \text{with probability } 1 - p_{it} \\
\mathrm{Beta}(\mu_{it}\phi,\ (1-\mu_{it})\phi) & \text{with probability } p_{it}
\end{cases}
}
}
\note{
A dedicated \code{\link{plot}} function is provided for objects of class \code{"zibSim"}.
}
\examples{
n = 1000
t = 3
X <- as.matrix(c(rep(0, n/2 * t), rep(1, n/2 * t)))

y1 <- zibSim(n = n, t = t, a = 0.5, b = 0.8, sigma1 = 2.5, sigma2 = 0.5,
phi = 7.2, alpha = 0.5, beta = -0.5, X = X, Z = X, seed = 6874)

y2 <- zibSim(n = n, t = t, a = 0.1, b = 0.9, sigma1 = 3.2, sigma2 = 2.1,
phi = 3.2, alpha = -0.5, beta = -0.5, X = X, Z = X, seed = 6874)


}
\seealso{
\code{\link{plot}}
}
